Factors out TobinTransfer function
EVM should not be aware of tobin transfer. From EVM's perspective; it's just a
transfer; the specific of what a transfer entails it not for the EVM to know

With this commit, `TobinTransfer` function is removed from the EVM. Instead,
the evm relies on EVM.Context.Transfer() function.

Now, the EVM.Context.Tranfer() function implementation is now actually the tobin
transfer. For this to work, context.Tranfer function requires the EVM as parameter

Additionally, the commit removes the uneeded `gas` parameter and the unnecessary
`leftGas` and `error` return parameters. The reasoon is that gas was actually not used
and the function never ended with error. This, in fact, makes the function share the
structure of the former context.Transfer() function.
