+++ core/vm/context.go
@@ -22,8 +22,10 @@ import (
 	"github.com/celo-org/celo-blockchain/common"
 	"github.com/celo-org/celo-blockchain/consensus"
 	"github.com/celo-org/celo-blockchain/consensus/istanbul"
+	"github.com/celo-org/celo-blockchain/contracts"
 	"github.com/celo-org/celo-blockchain/core/state"
 	"github.com/celo-org/celo-blockchain/core/types"
+	"github.com/celo-org/celo-blockchain/log"
 	"github.com/celo-org/celo-blockchain/params"
 )
 
@@ -88,7 +90,7 @@ func NewEVMContext(msg Message, header *types.Header, chain ChainContext, txFeeR
 
 	ctx := Context{
 		CanTransfer: CanTransfer,
-		Transfer:    Transfer,
+		Transfer:    TobinTransfer,
 		GetHash:     GetHashFn(header, chain),
 		VerifySeal:  VerifySealFn(header, chain),
 		Origin:      msg.From(),
@@ -174,4 +176,30 @@ func VerifySealFn(ref *types.Header, chain ChainContext) func(*types.Header) boo
 		// Submit the header to the engine's seal verification function.
 		return chain.Engine().VerifySeal(nil, header) == nil
 	}
+}
+
+// TobinTransfer performs a transfer that may take a tax from the sent amount and give it to the reserve.
+// If the calculation or transfer of the tax amount fails for any reason, the regular transfer goes ahead.
+func TobinTransfer(evm *EVM, sender, recipient common.Address, amount *big.Int) {
+	// Run only primary evm.Call() with tracer
+	if evm.GetDebug() {
+		evm.SetDebug(false)
+		defer func() { evm.SetDebug(true) }()
+	}
+
+	if amount.Cmp(big.NewInt(0)) != 0 {
+		tax, taxCollector, err := contracts.ComputeTobinTax(NewCaller(evm), sender, amount)
+		if err == nil {
+			Transfer(evm.StateDB, sender, recipient, new(big.Int).Sub(amount, tax))
+			Transfer(evm.StateDB, sender, taxCollector, tax)
+			return
+		} else {
+			log.Error("Failed to get tobin tax", "error", err)
+		}
+	}
+
+	// Complete a normal transfer if the amount is 0 or the tobin tax value is unable to be fetched and parsed.
+	// We transfer even when the amount is 0 because state trie clearing [EIP161] is necessary at the end of a transaction
+	Transfer(evm.StateDB, sender, recipient, amount)
+	return
 }
