+++ core/vm/gas_table_test.go
@@ -89,7 +89,7 @@ func TestEIP2200(t *testing.T) {
 
 		vmctx := Context{
 			CanTransfer: func(StateDB, common.Address, *big.Int) bool { return true },
-			Transfer:    func(StateDB, common.Address, common.Address, *big.Int) {},
+			Transfer:    func(*EVM, common.Address, common.Address, *big.Int) {},
 		}
 		vmenv := NewEVM(vmctx, statedb, params.IstanbulTestChainConfig, Config{ExtraEips: []int{2200}})
